move lib up for visibility
