+++ src/index.js
@@ -1,4 +1,4 @@
-import lib from './lib/index.js'
+import lib from './libs.js'
 import components from './components/index.js'
 
 import '../styles/index.css'
